Deploy WitNote as original desktop app using noVNC
- Set up a virtual desktop environment (Xvfb, Fluxbox, x11vnc, noVNC).
- Build and install the application as a .deb package inside Docker.
- Implement a FastAPI health/info server for Hugging Face Space monitoring.
- Configure Nginx as a gateway to serve noVNC and the API on port 7860.
- Update GitHub Actions to push all necessary configuration files.
- Ensure non-root user execution and security hardening.
