+++ index.html
@@ -140,7 +140,6 @@
     </div>
   </div>
 
-  <script src="/web-bridge.js"></script>
   <script>
     (function () {
       // Check if this is a refresh (sessionStorage persists on refresh, clears on close)
