+++ nginx.conf
@@ -1,0 +1,38 @@
+daemon off;
+error_log /tmp/nginx_error.log;
+pid /tmp/nginx.pid;
+
+events {
+    worker_connections 1024;
+}
+
+http {
+    access_log /tmp/nginx_access.log;
+    client_body_temp_path /tmp/client_body;
+    proxy_temp_path /tmp/proxy_temp;
+    fastcgi_temp_path /tmp/fastcgi_temp;
+    uwsgi_temp_path /tmp/uwsgi_temp;
+    scgi_temp_path /tmp/scgi_temp;
+
+    server {
+        listen 7860;
+
+        location /health {
+            proxy_pass http://127.0.0.1:7861/health;
+        }
+
+        location /api/ {
+            proxy_pass http://127.0.0.1:7861/api/;
+        }
+
+        location / {
+            proxy_pass http://127.0.0.1:6080/;
+            proxy_http_version 1.1;
+            proxy_set_header Upgrade $http_upgrade;
+            proxy_set_header Connection "Upgrade";
+            proxy_set_header Host $host;
+            proxy_read_timeout 61s;
+            proxy_buffering off;
+        }
+    }
+}
